+++ com.unity.render-pipelines.core/Editor/BuildProcessors/SRPShaderStripper.cs
@@ -1,6 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
 using UnityEngine;
-using UnityEngine.Rendering;
 
 namespace UnityEditor.Rendering
 {